Move privete headers to internal folder
