+++ bindings/c/include/khiva_c/internal/util.h
