+++ bindings/c/src/dimensionality.cpp
@@ -8,7 +8,7 @@
 #include <khiva/dimensionality.h>
 #include <khiva_c/dimensionality.h>
 #include <iostream>
-#include <khiva_c/util.h>
+#include <khiva_c/internal/util.h>
 
 void paa(khiva_array *a, int *bins, khiva_array *result, int *error_code, char *error_message) {
     try {
