+++ bindings/c/src/normalization.cpp
@@ -7,7 +7,7 @@
 #include <arrayfire.h>
 #include <khiva/normalization.h>
 #include <khiva_c/normalization.h>
-#include <khiva_c/util.h>
+#include <khiva_c/internal/util.h>
 
 void decimal_scaling_norm(khiva_array *tss, khiva_array *result, int *error_code, char *error_message) {
     try {
