+++ bindings/c/src/polynomial.cpp
@@ -9,7 +9,7 @@
 #include <arrayfire.h>
 #include <khiva/polynomial.h>
 #include <khiva_c/polynomial.h>
-#include <khiva_c/util.h>
+#include <khiva_c/internal/util.h>
 
 void polyfit(khiva_array *x, khiva_array *y, int *deg, khiva_array *result, int *error_code, char *error_message) {
     try {
